feat(json-schema): don't throw by default in removeAdditional
This changes the behavior so that by default removeAdditional will not throw
when it encounters a validation error. Instead it will log a message to a
provided 'errorLogger' and return the invalid object. Additional properties
that were not allowed in the schema will still be removed. The previous
behavior where removeAdditional would throw on validation errors can be
accomplished by setting the 'strict' option to true.